requirements.txt
sentence-transformers>=3.0.0
torch>=2.2.0
pandas>=2.2.0
pyarrow>=15.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
PyYAML>=6.0.0

